+++ sdk/appcenter-distribute/src/main/java/com/microsoft/appcenter/distribute/download/ReleaseDownloader.java
@@ -73,9 +73,10 @@ public interface ReleaseDownloader {
          * Called when the downloading is completed.
          *
          * @param downloadId downloadId of downloaded file.
+         * @param totalSize total size of downloaded file.
          */
         @WorkerThread
-        void onComplete(@NonNull Long downloadId);
+        void onComplete(@NonNull long downloadId, long totalSize);
 
         /**
          * Called when an error occurs during the downloading.