+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 * **[Feature]** Remove the download manager task if the download doesn't start within 10 seconds.
 * **[Feature]** Replace installing a new release using the deprecated intent action [ACTION_INSTALL_PACKAGE](https://developer.android.com/reference/android/content/Intent#ACTION_INSTALL_PACKAGE) with the `PackageInstaller` API.
+* **[Feature]** Add sumcheck on the downloaded file before starting the install process.
 
 ___